+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - [ ] Use FileSystemWatcher (#28)
 - [x] Universal Windows Apps support
 - [x] Globbing pattern support (#46)
+- [x] Added ignore file pattern (#54)
+  - InputFiles starting with `!` is ignored
 - [x] Support for files with spaces in the path (#55)
 
 Features that have a checkmark are complete and available for